Intercambio de datos de logico listo

+++ Midgard/DataAccess/jsonwritercpp.h
@@ -11,7 +11,7 @@ class jsonWriterCpp
 public:
     //Logic
 
-    string writeMap(Vector<short> pMatriz);
+    string writeMap(Vector<int> pMatriz);
     string writeFamily(int pFather, int pMother, int pFitness);
     string writeFight(int pPopulation1, int pPopulation2);
